Implement fake images

+++ server/src/controllers/post.controllers/remove.post.by.id.ts
@@ -27,14 +27,17 @@ export const remove = async (req: Request, res: Response, next: NextFunction): P
             throw new Error(`Unauthorized attempt to delete post ${postId}.`);
         }
 
-        uploadImage(awsConfig).s3.deleteObject({
-            Bucket: awsConfig.bucket,
-            Key: delPost.imgPath.split('/').pop(),
-        }, (error: Error, data: any): void => {
-            if (error) {
-                throw new Error(error.message);
-            }
-        });
+        if (delPost.imgPath.match(/amazonaws.com/)) { // do not delete test DB photos
+            uploadImage(awsConfig).s3.deleteObject({
+                Bucket: awsConfig.bucket,
+                Key: delPost.imgPath.split('/').pop(),
+            }, (error: Error, data: any): void => {
+                if (error) {
+                    throw new Error(error.message);
+                }
+            });
+        }
+
         res.json({message: 'Post was successfully deleted', delPost});
     } catch (e) {
         next({message: 'Couldn\'t delete post', status: 500});
